+++ bonsai/common/state_to_proto.py
@@ -11,10 +11,15 @@ class SimStateException(Exception):
 
 def build_luminance_from_state(field_name, proto_msg, luminance):
     """ This function sets a luminance datum onto a protobuf message """
-    lum_attr = getattr(proto_msg, field_name)
-    lum_attr.width = luminance.width
-    lum_attr.height = luminance.height
-    lum_attr.pixels = luminance.pixels
+    if luminance.__class__.__name__ == 'Luminance':
+        lum_attr = getattr(proto_msg, field_name)
+        lum_attr.width = luminance.width
+        lum_attr.height = luminance.height
+        lum_attr.pixels = luminance.pixels
+    else:
+        raise SimStateException(
+            'Expected bonsai.inkling_types.Luminance Got {}'
+            .format(luminance.__class__))
 
 
 # inkling_type_tensor_handler maps inkling types by name to handlers built
@@ -79,10 +84,10 @@ def convert_state_to_proto(state_msg, state):
             value = state[field.name]
         except KeyError:
             raise SimStateException(
-                'The inkling file specifies a field named "{}" which was not found in the '
-                'SimState. Please check the inkling file state schema and the return value '
-                'from get_state().'.format(field.name))
-
+                'The inkling file specifies a field named "{}" which was not '
+                'found in the SimState. Please check the inkling file state '
+                'schema and the return value from get_state().'
+                .format(field.name))
         # If the field is a message, assume it is Luminance.
         if is_proto_type_embedded_message(field):
             build_proto_from_embedded_type(